Move compiler main into a `compile` module.
