+++ src/scala/org/pantsbuild/zinc/analysis/AnalysisMap.scala
@@ -12,15 +12,13 @@ import java.util.Optional
 import scala.compat.java8.OptionConverters._
 
 import sbt.internal.inc.{Analysis, AnalysisMappersAdapter, AnalysisStore, CompanionsStore, Mapper, FileBasedStore, Locate}
-import sbt.io.{IO, Using}
 import sbt.util.Logger
 import xsbti.api.Companions
-import xsbti.compile.{CompileAnalysis, DefinesClass, MiniSetup, PerClasspathEntryLookup, PreviousResult}
+import xsbti.compile.{CompileAnalysis, DefinesClass, MiniSetup, PerClasspathEntryLookup}
 
 import org.pantsbuild.zinc.cache.Cache.Implicits
 import org.pantsbuild.zinc.cache.{Cache, FileFPrint}
 import org.pantsbuild.zinc.util.Util
-import org.pantsbuild.zinc.options.Settings
 
 /**
  * A facade around the analysis cache to:
@@ -90,37 +88,7 @@ class AnalysisMap private[AnalysisMap] (
        analysisLocations.get(classpathEntry).flatMap(cacheLookup)
   }
 
-  /**
-   * Load the analysis for the destination, creating it if necessary.
-   */
-  def loadDestinationAnalysis(
-    settings: Settings,
-    log: Logger
-  ): (AnalysisStore, PreviousResult) = {
-    def load() = {
-      val analysisStore = cachedStore(settings.cacheFile)
-      analysisStore.get() match {
-        case Some((a, s)) => (analysisStore, Some(a), Some(s))
-        case _ => (analysisStore, None, None)
-      }
-    }
-
-    // Try loading, and optionally remove/retry on failure.
-    val (analysisStore, previousAnalysis, previousSetup) =
-      try {
-        load()
-      } catch {
-        case e: Throwable if settings.analysis.clearInvalid =>
-          // Remove the corrupted analysis and output directory.
-          log.warn(s"Failed to load analysis from ${settings.cacheFile} ($e): will execute a clean compile.")
-          IO.delete(settings.cacheFile)
-          IO.delete(settings.classesDirectory)
-          load()
-      }
-    (analysisStore, new PreviousResult(previousAnalysis.asJava, previousSetup.asJava))
-  }
-
-  private def cachedStore(cacheFile: File): AnalysisStore =
+  def cachedStore(cacheFile: File): AnalysisStore =
     AnalysisStore.sync(
       new AnalysisStore {
         val fileStore = mkFileBasedStore(cacheFile)
