--- src/scala/org/pantsbuild/zinc/logging/BUILD
@@ -1,14 +1,0 @@
-scala_library(
-  provides=scala_artifact(
-    org='org.pantsbuild',
-    name='zinc-logging',
-    repo=public,
-    publication_metadata=pants_library('The SBT incremental compiler for nailgun')
-  ),
-  dependencies=[
-    '3rdparty/jvm/org/scala-sbt:util-logging',
-    '3rdparty/jvm/org/scala-sbt:zinc',
-  ],
-  strict_deps=True,
-  platform='java8',
-)
