+++ src/scala/org/pantsbuild/zinc/options/BUILD
@@ -5,10 +5,6 @@ scala_library(
     repo=public,
     publication_metadata=pants_library('The SBT incremental compiler for nailgun')
   ),
-  dependencies=[
-    '3rdparty/jvm/org/scala-lang/modules:scala-java8-compat',
-    '3rdparty/jvm/org/scala-sbt:zinc',
-  ],
   strict_deps=True,
   platform='java8',
 )
