+++ tests/scala/org/pantsbuild/zinc/analysis/AnalysisMapSpec.scala
@@ -1,7 +1,7 @@
 // Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-package org.pantsbuild.zinc
+package org.pantsbuild.zinc.analysis
 
 import sbt.io.IO
 import sbt.internal.util.{
@@ -19,7 +19,7 @@ class AnalysisMapSpec extends WordSpec with MustMatchers {
   "AnalysisMap" should {
     "succeed for empty analysis" in {
       IO.withTemporaryDirectory { classpathEntry =>
-        val am = AnalysisMap.create(Map(), ConsoleLogger(ConsoleOut.systemOut))
+        val am = AnalysisMap.create(Map(), Map(), ConsoleLogger(ConsoleOut.systemOut))
         val dc = am.getPCELookup.definesClass(classpathEntry)
         dc("NonExistent.class") must be(false)
       }
