--- tests/scala/org/pantsbuild/zinc/logging/BUILD
@@ -1,11 +1,0 @@
-# Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
-# Licensed under the Apache License, Version 2.0 (see LICENSE).
-
-junit_tests(
-  dependencies=[
-    '3rdparty:guava',
-    '3rdparty:junit',
-    '3rdparty:scalatest',
-    'src/scala/org/pantsbuild/zinc/logging',
-  ],
-)
